Remove externals from webpack dist configuration (#13)
* Remove externals from webpack dist configuration

This configuration causes an error resolving Vue instance when building on linux OS.

On Windows OS this error does not occur because of (I think) his non-case sensitive file system.

* Add again the externals webpack configuration

The fix was not delete the externals but change 'Vue' for 'vue'. This allows make the build without include the vue library.